Move direct award models (and related tests) into separate modules.
 - a by-product of this is that the main app models have also moved, but
   are still accessible via `app.model`

https://trello.com/c/pbJiKaVl/625-api-for-direct-award-projects
