+++ app/models/main.py
@@ -1,15 +1,17 @@
+# TODO split this file into per-functional-area modules
+
 import random
 from datetime import datetime
-import re
 
+import re
+import sqlalchemy.dialects.postgresql
 from flask import current_app
 from flask_sqlalchemy import BaseQuery
 from six import string_types, iteritems
-
+from sqlalchemy import Sequence
 from sqlalchemy import asc, desc
 from sqlalchemy import func
 from sqlalchemy.dialects.postgresql import INTERVAL
-import sqlalchemy.dialects.postgresql
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import validates, backref, mapper
@@ -22,15 +24,13 @@ from sqlalchemy.sql.expression import (
     and_ as sql_and,
 )
 from sqlalchemy.types import String
-from sqlalchemy import Sequence
 from sqlalchemy_utils import generic_relationship
 
-from dmutils.formats import DATETIME_FORMAT, DATE_FORMAT
 from dmutils.dates import get_publishing_dates
-
-from . import db
-from .utils import link, url_for, strip_whitespace_from_data, drop_foreign_fields, purge_nulls_from_data
-from .validation import is_valid_service_id, is_valid_buyer_email, get_validation_errors
+from dmutils.formats import DATETIME_FORMAT, DATE_FORMAT
+from .. import db
+from ..utils import link, url_for, strip_whitespace_from_data, drop_foreign_fields, purge_nulls_from_data
+from ..validation import is_valid_service_id, is_valid_buyer_email, get_validation_errors
 
 
 class JSON(sqlalchemy.dialects.postgresql.JSON):
@@ -1646,99 +1646,6 @@ class BriefClarificationQuestion(db.Model):
             "answer": self.answer,
             "publishedAt": self.published_at.strftime(DATETIME_FORMAT),
         }
-
-
-class DirectAwardProject(db.Model):
-    __tablename__ = 'direct_award_projects'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(255), nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
-    locked_at = db.Column(db.DateTime, nullable=True)  # When the project's active search/es are final and cannot change
-    active = db.Column(db.Boolean, default=True, nullable=False)
-
-    users = db.relationship(User, secondary='direct_award_project_users', order_by=lambda: DirectAwardProjectUser.id)
-
-    def serialize(self):
-        return {
-            "id": self.id,
-            "name": self.name,
-            "created_at": self.created_at.strftime(DATETIME_FORMAT),
-            "locked_at": self.locked_at.strftime(DATETIME_FORMAT) if self.locked_at is not None else None,
-            "active": self.active
-        }
-
-    @validates('name')
-    def _assert_active_project(self, key, value):
-        if self.locked_at:
-            raise ValidationError('Cannot change project name after locking it ({})'.format(self.id))
-
-        return value
-
-
-class DirectAwardProjectUser(db.Model):
-    __tablename__ = 'direct_award_project_users'
-
-    id = db.Column(db.Integer, primary_key=True)
-    project_id = db.Column(db.Integer, db.ForeignKey('direct_award_projects.id'), index=True, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), index=True, nullable=False)
-
-
-class DirectAwardSearch(db.Model):
-    __tablename__ = 'direct_award_searches'
-
-    id = db.Column(db.Integer, primary_key=True)
-    created_by = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    project_id = db.Column(db.Integer, db.ForeignKey('direct_award_projects.id'), nullable=False, index=True)
-    created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)  # When this search was first saved.
-    searched_at = db.Column(db.DateTime, nullable=True)  # When this search was run for the project services download.
-    search_url = db.Column(db.Text, nullable=False)  # Search API URL to hit when retrieving services.
-    active = db.Column(db.Boolean, default=True, nullable=False, index=True)  # Is the 'current' search for the project
-
-    # Access archived_services relating to this search. Should only have related services after the project
-    # has been downloaded, and this is the search for that project that was chosen as the search to longlist/shortlist
-    # against.
-    # Note: an archived service is not a service from an old framework, as one might think. An ArchivedService is
-    # created when an attribute on a service is changed. By storing the ArchivedService id we are referencing a service
-    # in a given state at a specific time. We can still access the live version of the service through the Archived
-    # Service, but by storing this we have the flexibility to show either the live state or the state at the time
-    # the user ran the search.
-    archived_services = db.relationship(ArchivedService, secondary='direct_award_search_result_entries',
-                                        order_by=lambda: DirectAwardSearchResultEntry.id)
-
-    project = db.relationship(DirectAwardProject)
-    user = db.relationship(User)
-
-    # Partial index on project_id,active==1. Enforces only one active search per project at a time.
-    __table_args__ = (db.Index('idx_project_id_active', project_id, active, unique=True,
-                               postgresql_where=db.Column('active')),)
-
-    def serialize(self):
-        return {
-            "id": self.id,
-            "created_by": self.created_by,
-            "project_id": self.project_id,
-            "created_at": self.created_at.strftime(DATETIME_FORMAT),
-            "searched_at": self.searched_at.strftime(DATETIME_FORMAT) if self.searched_at is not None else None,
-            "search_url": self.search_url,
-            "active": self.active
-        }
-
-    @validates('id', 'created_by', 'project_id', 'created_at', 'searched_at', 'search_url', 'active')
-    def _assert_active_project(self, key, value):
-        if self.project and self.project.locked_at:
-            raise ValidationError('Cannot change attributes of a search under a project ({}) that '
-                                  'has been locked.'.format(self.project_id))
-
-        return value
-
-
-class DirectAwardSearchResultEntry(db.Model):
-    __tablename__ = 'direct_award_search_result_entries'
-
-    id = db.Column(db.Integer, primary_key=True)
-    search_id = db.Column(db.Integer, db.ForeignKey('direct_award_searches.id'), index=True, nullable=False)
-    archived_service_id = db.Column(db.Integer, db.ForeignKey('archived_services.id'), index=True, nullable=False)
 
 
 # Index for .last_for_object queries. Without a composite index the
